reworked api import from star imports to normal imports, added child picture upload and return end-points
