+++ server/backbone/decorators.py
@@ -1,8 +1,8 @@
 import inspect
-from functools import wraps
 
 from flask import abort, request, json
 from flask_login import current_user, login_required
+from functools import wraps
 
 from .exceptions import BackboneException
 from .models import Parent, Child
